requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pdfplumber>=0.10.0
python-docx>=1.0.0
pandas>=2.0.0
